update: extract from page path generator
