+++ frontend/src/components/NodeExplorer/index.tsx
@@ -4,7 +4,7 @@ import type { IPLDNodeData } from '@autonomys/auto-dag-data';
 import { ChevronDownIcon, ChevronRightIcon } from 'lucide-react';
 import { FC, useState } from 'react';
 import { InternalLink } from '../common/InternalLink';
-import { getFSPath } from '../../app/[chain]/drive/fs/[cid]/page';
+import { getFSPath } from '../../views/FileSystem';
 import { useNetwork } from '../../contexts/network';
 
 interface NodeExplorerProps {
